# spec/openapi.yaml
openapi: 3.0.0
info:
  title: Knobel Manager API
  version: 1.0.0
  description: API for Knobel Manager Service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
paths:
...
  /games/{gameID}/teams:
    parameters:
      - name: gameID
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: createTeam
      tags: [ Teams ]
      summary: Create a team
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamsRequest'
      responses:
        '201':
          description: Team created
          headers:
            Location:
              description: URL of the created team
              schema:
                type: string
                example: /games/1/teams/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request
        '403':
          description: Not owner of the game
        '404':
          description: Game not found
  /games/{gameID}/teams/{teamID}:
    parameters:
      - name: gameID
        in: path
        required: true
        schema:
          type: integer
      - name: teamID
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateTeam
      tags: [ Teams ]
      summary: Update a team
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamsRequest'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request
        '403':
          description: Not owner of the game
        '404':
          description: Team or game not found
    delete:
      operationId: deleteTeam
      tags: [ Teams ]
      summary: Delete a team
      security:
        - bearerAuth: [ ]
      responses:
        '204':
          description: Team deleted
        '403':
          description: Not owner of the game
        '404':
          description: Team or game not found
...
  /games/{gameID}/rounds/{roundNumber}/tables:
    parameters:
      - name: gameID
        in: path
        required: true
        schema:
          type: integer
      - name: roundNumber
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getTables
      tags: [ Tables ]
      summary: List tables for a round
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Tables found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TablesResponse'
        '400':
          description: Invalid path parameters
        '403':
          description: Not owner of the game
        '404':
          description: Game or round not found
  /games/{gameID}/rounds/{roundNumber}/tables/{tableNumber}:
    parameters:
      - name: gameID
        in: path
        required: true
        schema:
          type: integer
      - name: roundNumber
        in: path
        required: true
        schema:
          type: integer
      - name: tableNumber
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getTable
      tags: [ Tables ]
      summary: Get a table by number in round
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Table found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          description: Invalid path parameters
        '403':
          description: Not owner of the game
        '404':
          description: Round or table not found
  /games/{gameID}/rounds/{roundNumber}/tables/{tableNumber}/scores:
    parameters:
      - name: gameID
        in: path
        required: true
        schema:
          type: integer
      - name: roundNumber
        in: path
        required: true
        schema:
          type: integer
      - name: tableNumber
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateScores
      tags: [ Scores ]
      summary: Update scores for a table
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoresRequest'
      responses:
        '200':
          description: Updated game with new scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          description: Invalid request body or path params
        '403':
          description: Not owner of the game
        '404':
          description: Game, round, or table not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GameOwner:
      type: object
      properties:
        gameID:
          type: integer
          example: 1
        ownerSub:
          type: string
          example: sub-1
      required: [ gameID, ownerSub ]
    Player:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Player 1
        teamID:
          type: integer
          example: 1
      required: [ id, name, teamID ]
    Team:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Team A
        gameID:
          type: integer
          example: 1
        players:
          type: array
          items:
            $ref: '#/components/schemas/Player'
      required: [ id, name, gameID ]
    Table:
      type: object
      properties:
        id:
          type: integer
          example: 10
        tableNumber:
          type: integer
          example: 1
        roundID:
          type: integer
          example: 5
        players:
          type: array
          items:
            $ref: '#/components/schemas/Player'
        scores:
          type: array
          items:
            $ref: '#/components/schemas/Score'
      required: [ id, tableNumber, roundID ]
    Score:
      type: object
      properties:
        id:
          type: integer
          example: 100
        playerID:
          type: integer
          example: 1
        tableID:
          type: integer
          example: 10
        score:
          type: integer
          example: 6
      required: [ id, playerID, tableID, score ]
    GameRound:
      type: object
      properties:
        id:
          type: integer
        gameID:
          type: integer
        roundNumber:
          type: integer
        status:
          $ref: '#/components/schemas/RoundStatus'
        tables:
          type: array
          items:
            $ref: '#/components/schemas/Table'
      required: [ id, gameID, roundNumber, status ]
    Game:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Game 1
        teamSize:
          type: integer
          example: 4
        tableSize:
          type: integer
          example: 4
        numberOfRounds:
          type: integer
          example: 2
        status:
          type: string
          enum: [ setup, in_progress,completed ]
          example: setup
        owners:
          type: array
          items:
            $ref: '#/components/schemas/GameOwner'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/Team'
        rounds:
          type: array
          items:
            $ref: '#/components/schemas/GameRound'
      required:
        - id
        - name
        - teamSize
        - tableSize
        - numberOfRounds
        - status
        - owners
...
    PlayersRequest:
      type: object
      properties:
        name:
          type: string
      required: [ name ]
...
    TeamsRequest:
      type: object
      properties:
        name:
          type: string
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayersRequest'
      required: [ name ]
    ScoresRequest:
      type: object
      properties:
        scores:
          type: array
          items:
            type: object
            properties:
              playerID:
                type: integer
              score:
                type: integer
            required: [ playerID, score ]
      required: [ scores ]
...
    GameResponse:
      type: object
      properties:
        game:
          $ref: '#/components/schemas/Game'
      required: [ game ]
    TeamResponse:
      type: object
      properties:
        team:
          $ref: '#/components/schemas/Team'
      required: [ team ]
    TablesResponse:
      type: object
      properties:
        tables:
          type: array
          items:
            $ref: '#/components/schemas/Table'
      required: [ tables ]
...
    RoundStatus:
      type: string
      enum: [ setup, in_progress, completed ]
      example: in_progress